Fix various tests; removed inspection to determine if strategy is stochastic due to inconsistency and favoring explicit over implicit
